# feat: optimization of rayon 's usage