@flaneur2020 2024-03-19
> Evething looks fine?

sorry, we'd prefer to focus on optimizing the larger models (> 3B parameters) to achieve performance as fast as llama.cpp. multithreaded parallelism in Matmul is still a must unless benchmarks show that we can generate 100 tokens on a 3B model faster than llama.cpp.