@xiaguan 2024-03-19
Is any link to download a biggger model？