@MrCroxx 2024-03-19
> Is any link to download a biggger model？

Try this: https://github.com/crabml/crabml/issues/110#issuecomment-1999007793