@flaneur2020 2024-03-19
i also have a small benchmark script might be useful: https://gist.github.com/flaneur2020/27a384e8a6eae8963491c0bbf6bb9033 

(which is generated by GPT4 though)

(i suspect that llama.cpp generates 99 tokens while specifying `-n 100`, maybe specifying `--steps 99` might be more fair on the benchmark)