@xiaguan 2024-03-19
For big model ,it arg is big enough (3000,8000, always bigger than 1000).
test command
```
cargo run --release -- -m ./testdata/open-llama-3b-v2-instruct.q8_0.gguf  'capatain' --steps 100 -t 0 -T 2
```
`main` branch
```
4.924048 tokens/s, 2 threads
7.0392613 tokens/s, 16 threads
```
perf stat for 16 threads
```
        219,423.24 msec task-clock                       #   13.403 CPUs utilized             
           275,716      context-switches                 #    1.257 K/sec                     
               661      cpu-migrations                   #    3.012 /sec                      
           152,679      page-faults                      #  695.820 /sec                      
   717,952,274,389      cycles                           #    3.272 GHz                         (83.31%)
    12,972,192,592      stalled-cycles-frontend          #    1.81% frontend cycles idle        (83.32%)
   583,612,087,633      stalled-cycles-backend           #   81.29% backend cycles idle         (83.29%)
   357,562,718,579      instructions                     #    0.50  insn per cycle            
                                                  #    1.63  stalled cycles per insn     (83.40%)
    25,543,858,870      branches                         #  116.414 M/sec                       (83.41%)
       837,256,315      branch-misses                    #    3.28% of all branches             (83.26%)

      16.371488721 seconds time elapsed

     205.882895000 seconds user
      12.995128000 seconds sys
```
my branch
threads = 2 
```
5.0425935 tokens/s, 2 threads
```
threads = 8
```
6.406336 tokens/s, 8 threads
```
threads = 16
```
6.714745 tokens/s, 16 threads
```
perf stat for 16 threads
```
        199,575.54 msec task-clock                       #   11.612 CPUs utilized             
           512,709      context-switches                 #    2.569 K/sec                     
               904      cpu-migrations                   #    4.530 /sec                      
           146,169      page-faults                      #  732.399 /sec                      
   557,562,847,503      cycles                           #    2.794 GHz                         (83.29%)
     9,273,378,419      stalled-cycles-frontend          #    1.66% frontend cycles idle        (83.47%)
   412,121,158,296      stalled-cycles-backend           #   73.91% backend cycles idle         (83.22%)
   365,615,343,293      instructions                     #    0.66  insn per cycle            
                                                  #    1.13  stalled cycles per insn     (83.31%)
    29,089,326,383      branches                         #  145.756 M/sec                       (83.34%)
     1,132,059,360      branch-misses                    #    3.89% of all branches             (83.37%)

      17.187602040 seconds time elapsed

     178.684921000 seconds user
      20.363169000 seconds sys
```
My cpu has 24 core , but when I set thread num to a big one. It will panic on my branch.
```
loaded model: 10ms
capatain春thread 'main' panicked at crabml-core/src/tokenizer/bpe.rs:71:55:
called `Result::unwrap()` on an `Err` value: FromUtf8Error { bytes: [231], error: Utf8Error { valid_up_to: 0, error_len: None } }
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

```