@flaneur2020 2024-03-19
> My cpu has 24 core , but when I set thread num to a big one. It will panic on my branch.
> loaded model: 10ms
> capatain春thread 'main' panicked at crabml-core/src/tokenizer/bpe.rs:71:55:
> called `Result::unwrap()` on an `Err` value: FromUtf8Error { bytes: [231], error: Utf8Error { valid_up_to: 0, error_len: None } }
> note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace

I believe that the current bpe tokenizer has some bugs on handling utf-8 >_<