@xiaguan 2024-03-21
Seems like rayon isn't the bottleneck with the larger models, right? I'm gonna go ahead and close this PR for now.